PUBDEV-5742 WIP. Added target encoding params into AUtoMLBuildSpec. Added FixedTEParamsStrategy to specify single target encoding params for whole automl process.

+++ h2o-automl/src/main/java/ai/h2o/automl/AutoMLTargetEncodingAssistant.java
@@ -2,7 +2,10 @@ package ai.h2o.automl;
 
 import ai.h2o.automl.targetencoding.BlendingParams;
 import ai.h2o.automl.targetencoding.TargetEncoder;
+import ai.h2o.automl.targetencoding.TargetEncodingParams;
 import ai.h2o.automl.targetencoding.strategy.TEApplicationStrategy;
+import ai.h2o.automl.targetencoding.strategy.TEParamsSelectionStrategy;
+import water.Iced;
 import water.fvec.Frame;
 import water.fvec.Vec;
 
@@ -11,7 +14,7 @@ import java.util.Map;
 /**
  * Perform TargetEncoding for AutoML process based on a strategy of searching best parameters.
  */
-class AutoMLTargetEncodingAssistant {
+class AutoMLTargetEncodingAssistant{
 
   private Frame _trainingFrame;   
   private Frame _validationFrame; 
@@ -20,6 +23,7 @@ class AutoMLTargetEncodingAssistant {
   private Vec _foldColumn;
   private AutoMLBuildSpec _buildSpec;
 
+  private TEParamsSelectionStrategy _teParamsSelectionStrategy;
   private TEApplicationStrategy _applicationStrategy;
 
   AutoMLTargetEncodingAssistant(Frame trainingFrame,
@@ -28,6 +32,7 @@ class AutoMLTargetEncodingAssistant {
                                 Vec responseColumn,
                                 Vec foldColumn,
                                 AutoMLBuildSpec buildSpec,
+                                TEParamsSelectionStrategy teParamsSelectionStrategy,
                                 TEApplicationStrategy applicationStrategy) {
     _trainingFrame = trainingFrame;
     _validationFrame = validationFrame;
@@ -35,56 +40,76 @@ class AutoMLTargetEncodingAssistant {
     _responseColumn = responseColumn;
     _foldColumn = foldColumn;
     _buildSpec = buildSpec;
+    _teParamsSelectionStrategy = teParamsSelectionStrategy;
     _applicationStrategy = applicationStrategy;
   }
 
   void performAutoTargetEncoding() {
+    
     String[] columnsToEncode = getApplicationStrategy().getColumnsToEncode();
+    
     if(columnsToEncode.length > 0) {
 
-      //TODO Either perform random grid search over parameters or introduce evolutionary selection algo
-      BlendingParams blendingParams = new BlendingParams(5, 1);
-      boolean withBlendedAvg = true;
-      boolean imputeNAsWithNewCategory = true;
+      TargetEncodingParams teParams = _teParamsSelectionStrategy.getBestParams();
+      BlendingParams blendingParams = teParams.getBlendingParams();
+      boolean withBlendedAvg = teParams.isWithBlendedAvg();
+      boolean imputeNAsWithNewCategory = teParams.isImputeNAsWithNewCategory();
       long seed = _buildSpec.build_control.stopping_criteria.seed(); // TODO make it a dedicated parameter for users to set
-      byte holdoutType = TargetEncoder.DataLeakageHandlingStrategy.KFold;
+      byte holdoutType = teParams.getHoldoutType();
 
       TargetEncoder tec = new TargetEncoder(columnsToEncode, blendingParams);
 
       String responseColumnName = _trainingFrame.name(_trainingFrame.find(_responseColumn));
-      String foldColumnName = _trainingFrame.name(_trainingFrame.find(_foldColumn));
 
-      Map<String, Frame> encodingMap = tec.prepareEncodingMap(_trainingFrame, responseColumnName, foldColumnName, imputeNAsWithNewCategory);
+      Map<String, Frame> encodingMap = null;
+      
       switch (holdoutType) {
         case TargetEncoder.DataLeakageHandlingStrategy.KFold:
 
-          Frame encodedTrainingFrame = tec.applyTargetEncoding(_trainingFrame, responseColumnName, encodingMap, holdoutType, foldColumnName, withBlendedAvg, imputeNAsWithNewCategory, seed);
+          encodingMap = tec.prepareEncodingMap(_trainingFrame, responseColumnName, getFoldColumnName(), imputeNAsWithNewCategory);
+
+          Frame encodedTrainingFrame = tec.applyTargetEncoding(_trainingFrame, responseColumnName, encodingMap, holdoutType, getFoldColumnName(), withBlendedAvg, imputeNAsWithNewCategory, seed);
           copyEncodedColumnsToDestinationFrame(columnsToEncode, encodedTrainingFrame, _trainingFrame);
 
           if(_validationFrame != null) {
-            Frame encodedValidationFrame = tec.applyTargetEncoding(_validationFrame, responseColumnName, encodingMap, TargetEncoder.DataLeakageHandlingStrategy.None, foldColumnName, withBlendedAvg, imputeNAsWithNewCategory, seed);
+            Frame encodedValidationFrame = tec.applyTargetEncoding(_validationFrame, responseColumnName, encodingMap, TargetEncoder.DataLeakageHandlingStrategy.None, getFoldColumnName(), withBlendedAvg, imputeNAsWithNewCategory, seed);
             copyEncodedColumnsToDestinationFrame(columnsToEncode, encodedValidationFrame, _validationFrame);
           }
           if(_leaderboardFrame != null) {
-            Frame encodedLeaderboardFrame = tec.applyTargetEncoding(_leaderboardFrame, responseColumnName, encodingMap, TargetEncoder.DataLeakageHandlingStrategy.None, foldColumnName, withBlendedAvg, imputeNAsWithNewCategory, seed);
+            Frame encodedLeaderboardFrame = tec.applyTargetEncoding(_leaderboardFrame, responseColumnName, encodingMap, TargetEncoder.DataLeakageHandlingStrategy.None, getFoldColumnName(), withBlendedAvg, imputeNAsWithNewCategory, seed);
             copyEncodedColumnsToDestinationFrame(columnsToEncode, encodedLeaderboardFrame, _leaderboardFrame);
           }
           break;
+          
         case TargetEncoder.DataLeakageHandlingStrategy.LeaveOneOut:
+          encodingMap = tec.prepareEncodingMap(_trainingFrame, responseColumnName, null);
+          
           Frame encodedTrainingFrameLOO = tec.applyTargetEncoding(_trainingFrame, responseColumnName, encodingMap, holdoutType, withBlendedAvg,  imputeNAsWithNewCategory,seed);
           copyEncodedColumnsToDestinationFrame(columnsToEncode, encodedTrainingFrameLOO, _trainingFrame);
 
-          // TODO: it is a duplicate to KFold's transformations. Consider to refactor.
           if(_validationFrame != null) {
-            Frame encodedValidationFrame = tec.applyTargetEncoding(_validationFrame, responseColumnName, encodingMap, TargetEncoder.DataLeakageHandlingStrategy.None, foldColumnName, withBlendedAvg, imputeNAsWithNewCategory, seed);
+            Frame encodedValidationFrame = tec.applyTargetEncoding(_validationFrame, responseColumnName, encodingMap, TargetEncoder.DataLeakageHandlingStrategy.None, withBlendedAvg, imputeNAsWithNewCategory, seed);
             copyEncodedColumnsToDestinationFrame(columnsToEncode, encodedValidationFrame, _validationFrame);
           }
           if(_leaderboardFrame != null) {
-            Frame encodedLeaderboardFrame = tec.applyTargetEncoding(_leaderboardFrame, responseColumnName, encodingMap, TargetEncoder.DataLeakageHandlingStrategy.None, foldColumnName, withBlendedAvg, imputeNAsWithNewCategory, seed);
+            Frame encodedLeaderboardFrame = tec.applyTargetEncoding(_leaderboardFrame, responseColumnName, encodingMap, TargetEncoder.DataLeakageHandlingStrategy.None, withBlendedAvg, imputeNAsWithNewCategory, seed);
             copyEncodedColumnsToDestinationFrame(columnsToEncode, encodedLeaderboardFrame, _leaderboardFrame);
           }
           break;
         case TargetEncoder.DataLeakageHandlingStrategy.None:
+          encodingMap = tec.prepareEncodingMap(_trainingFrame, responseColumnName, null);
+
+          Frame encodedTrainingFrameNone = tec.applyTargetEncoding(_trainingFrame, responseColumnName, encodingMap, holdoutType, withBlendedAvg,  imputeNAsWithNewCategory,seed);
+          copyEncodedColumnsToDestinationFrame(columnsToEncode, encodedTrainingFrameNone, _trainingFrame);
+
+          if(_validationFrame != null) {
+            Frame encodedValidationFrameNone = tec.applyTargetEncoding(_validationFrame, responseColumnName, encodingMap, TargetEncoder.DataLeakageHandlingStrategy.None, withBlendedAvg, imputeNAsWithNewCategory, seed);
+            copyEncodedColumnsToDestinationFrame(columnsToEncode, encodedValidationFrameNone, _validationFrame);
+          }
+          if(_leaderboardFrame != null) {
+            Frame encodedLeaderboardFrameNone = tec.applyTargetEncoding(_leaderboardFrame, responseColumnName, encodingMap, TargetEncoder.DataLeakageHandlingStrategy.None, withBlendedAvg, imputeNAsWithNewCategory, seed);
+            copyEncodedColumnsToDestinationFrame(columnsToEncode, encodedLeaderboardFrameNone, _leaderboardFrame);
+          }
       }
       encodingMapCleanUp(encodingMap);
     }
@@ -104,13 +129,20 @@ class AutoMLTargetEncodingAssistant {
   }
 
   private void encodingMapCleanUp(Map<String, Frame> encodingMap) {
-    for( Map.Entry<String, Frame> map : encodingMap.entrySet()) {
-      map.getValue().delete();
+    if(encodingMap != null) {
+      for (Map.Entry<String, Frame> map : encodingMap.entrySet()) {
+        map.getValue().delete();
+      }
     }
   }
 
   public TEApplicationStrategy getApplicationStrategy() {
     return _applicationStrategy;
+  }
+
+
+  public String getFoldColumnName() {
+    return _trainingFrame.name(_trainingFrame.find(_foldColumn));
   }
 
   public void setApplicationStrategy(TEApplicationStrategy applicationStrategy) {
